+++ convey/mail_draft.py
@@ -14,6 +14,10 @@ class MailDraft:
         self.template_file = get_path(filename)
         self.mail_file = Path(Config.get_cache_dir(), filename)
 
+    @classmethod
+    def init(cls, parser):
+        cls.parser = parser
+
     def get_mail_preview(self) -> str:
         l = self.get_envelope().preview().splitlines()
         s = "\n".join(l[:40])
@@ -55,18 +59,16 @@ class MailDraft:
 
             if not e._sender:  # XX this should be a publicly exposed method (and internal ._sender may change to ._from in the future)
                 e.from_(Config.get("email_from_name", "SMTP"))
-            if not e._message or not e._message.strip():
+            if not e.message() or not e.message().strip():
                 return "Missing body text. Try writing 'Subject: text', followed by a newline a text."
-            if not e._subject:
+            if not e.subject():
                 return "Missing subject. Try writing 'Subject: text', followed by a newline."
             if attachment:
                 if attachment.path and attachment.attach and Config.get('attach_files', 'SMTP', get=bool):
                     e.attach(attachment.path, "text/csv", attachment.parser.attachment_name)
 
                 if Config.is_testing():
-                    e._to.clear()
-                    e._cc.clear()
-                    e._bcc.clear()
+                    e.recipients(clear=True)
                     intended_to = attachment.mail
                     e.to(Config.get('testing_mail'))
                     e.message(f"This is testing mail only from Convey."
@@ -87,15 +89,18 @@ class MailDraft:
 
     def jinja(self, attachment: "Attachment"):
         def print_attachment():
-            """ Prints the attachment contents and prevent it to be attached. """
+            """ Prints the attachment contents and prevent it to be attached.
+                # XX may have header=False parameter to skip header.
+            """
             attachment.attach = False
             return attachment.path.read_text()
 
         def amount(count=2):
             """ Check if the attachment has at least count number of lines.
                 Header is not counted.
-                # XX skip header
             """
+            if self.parser.settings["header"]:
+                count += 1  # we skip header, so we must increment the count
             with attachment.path.open() as f:
                 for i, _ in enumerate(f):
                     if i + 1 == count:
@@ -103,23 +108,25 @@ class MailDraft:
             return False
 
         def row():
-            """ Generate attachment row by row """
-            # XX skip header?
+            """ Generate attachment row by row. Header is skipped. """
             with attachment.path.open() as f:
+                if self.parser.settings["header"]:
+                    next(f)
                 for line in f:
-                    yield line.strip().split(",")  # XXX correct delimiter
+                    yield line.strip().split(self.parser.settings["dialect"].delimiter)
 
         def joined(column: int, delimiter=", "):
             """ Return a column joined by delimiter  """
             return delimiter.join(r[column] for r in row())
 
         # Access first line fields
-        first_line = next(row())
+        first_row = next(row())
         try:
-            self.text = Template(self.text).render(first_line=first_line,
+            self.text = Template(self.text).render(first_row=first_row,
                                                    row=row,
                                                    joined=joined,
                                                    amount=amount,
                                                    print_attachment=print_attachment)
         except exceptions.TemplateError as e:
+            print(f"Template error: {e}")
             return False